Move random hodler to be Supplier, other updates from code review.
