+++ docs/apidiffs/current_vs_latest/opentelemetry-sdk-common.txt
@@ -1,8 +1,2 @@
 Comparing source compatibility of  against 
-+++  NEW INTERFACE: PUBLIC(+) ABSTRACT(+) io.opentelemetry.sdk.common.RandomHolder  (not serializable)
-	+++  CLASS FILE FORMAT VERSION: 52.0 <- n.a.
-	+++  NEW SUPERCLASS: java.lang.Object
-	+++  NEW METHOD: PUBLIC(+) STATIC(+) io.opentelemetry.sdk.common.RandomHolder androidFriendly()
-	+++  NEW METHOD: PUBLIC(+) STATIC(+) io.opentelemetry.sdk.common.RandomHolder create(java.util.Random)
-	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) java.util.Random getRandom()
-	+++  NEW METHOD: PUBLIC(+) STATIC(+) io.opentelemetry.sdk.common.RandomHolder platformDefault()
+No changes.
