+++ sdk/common/src/main/java/io/opentelemetry/sdk/internal/AndroidFriendlyRandomHolder.java
@@ -3,21 +3,22 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.common;
+package io.opentelemetry.sdk.internal;
 
 import java.util.Random;
+import java.util.function.Supplier;
 
 /**
- * {@link RandomHolder} instance that doesn't use {@link java.util.concurrent.ThreadLocalRandom},
+ * {@link RandomSupplier} instance that doesn't use {@link java.util.concurrent.ThreadLocalRandom},
  * which is broken on most versions of Android (it uses the same seed everytime it starts up).
  */
-enum AndroidFriendlyRandomHolder implements RandomHolder {
+enum AndroidFriendlyRandomHolder implements Supplier<Random> {
   INSTANCE;
 
   private static final Random random = new Random();
 
   @Override
-  public Random getRandom() {
+  public Random get() {
     return random;
   }
 }
