+++ sdk/common/src/main/java/io/opentelemetry/sdk/internal/RandomSupplier.java
@@ -3,17 +3,20 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.common;
+package io.opentelemetry.sdk.internal;
 
 import java.util.Random;
-import javax.annotation.concurrent.ThreadSafe;
+import java.util.concurrent.ThreadLocalRandom;
+import java.util.function.Supplier;
 
-/** Interface that pulls the most efficient random per-platform. */
-@ThreadSafe
-public interface RandomHolder {
-
-  /** Returns the random number generator to use. */
-  Random getRandom();
+/**
+ * Provides random number generater constructor utilities.
+ *
+ * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
+ * at any time.
+ */
+public final class RandomSupplier {
+  private RandomSupplier() {}
 
   /**
    * Returns the platform default for random number generation.
@@ -21,12 +24,12 @@ public interface RandomHolder {
    * <p>The underlying implementation attempts to use {@link java.util.concurrent.ThreadLocalRandom}
    * on platforms where this is the most efficient.
    */
-  static RandomHolder platformDefault() {
+  public static Supplier<Random> platformDefault() {
     // note: check borrowed from OkHttp's check for Android.
     if ("Dalvik".equals(System.getProperty("java.vm.name"))) {
       return androidFriendly();
     }
-    return ThreadLocalRandomHolder.INSTANCE;
+    return ThreadLocalRandom::current;
   }
 
   /**
@@ -35,16 +38,7 @@ public interface RandomHolder {
    * <p>On android, ThreadLocalRandom is instantiated with the same initial seed on all new threads
    * leading to poor randomness.
    */
-  static RandomHolder androidFriendly() {
+  public static Supplier<Random> androidFriendly() {
     return AndroidFriendlyRandomHolder.INSTANCE;
-  }
-
-  /**
-   * Returns a custom random number holder.
-   *
-   * @param random The implementation of random to use.
-   */
-  static RandomHolder create(Random random) {
-    return () -> random;
   }
 }
